got 3 node motif algo back to working version
