+++ py-raphtory/src/algorithms.rs
@@ -18,9 +18,11 @@ use raphtory::algorithms::directed_graph_density::directed_graph_density as dire
 use raphtory::algorithms::generic_taint::generic_taint as generic_taint_rs;
 use raphtory::algorithms::local_clustering_coefficient::local_clustering_coefficient as local_clustering_coefficient_rs;
 use raphtory::algorithms::local_triangle_count::local_triangle_count as local_triangle_count_rs;
-use raphtory::algorithms::motifs::three_node_local::global_temporal_three_node_motif as global_temporal_three_node_motif_rs;
-use raphtory::algorithms::motifs::three_node_local::global_temporal_three_node_motif_from_local as global_temporal_three_node_motif_from_local_rs;
-use raphtory::algorithms::motifs::three_node_local::temporal_three_node_motif as temporal_three_node_motif_rs;
+use raphtory::algorithms::motifs::three_node_broken::global_temporal_three_node_motif as global_temporal_three_node_motif_rs;
+use raphtory::algorithms::motifs::three_node_broken::global_temporal_three_node_motif_from_local as global_temporal_three_node_motif_from_local_rs;
+use raphtory::algorithms::motifs::three_node_broken::temporal_three_node_motif as temporal_three_node_motif_rs;
+use raphtory::algorithms::motifs::three_node_local::all_motifs_count as all_motifs_rs;
+use raphtory::algorithms::motifs::three_node_local::global_motifs_count as global_motifs_rs;
 use raphtory::algorithms::pagerank::unweighted_page_rank;
 use raphtory::algorithms::reciprocity::{
     all_local_reciprocity as all_local_reciprocity_rs, global_reciprocity as global_reciprocity_rs,
@@ -238,4 +240,11 @@ pub fn global_temporal_three_node_motif_from_local(
     counts: HashMap<String, Vec<usize>>,
 ) -> Vec<usize> {
     global_temporal_three_node_motif_from_local_rs(counts)
+}
+
+#[pyfunction]
+pub fn all_local_motifs(
+    g: &PyGraphView, delta: i64
+) -> HashMap<u64, Vec<usize>> {
+    all_motifs_rs(&g.graph, delta)
 }
