+++ python/src/lib.rs
@@ -50,6 +50,10 @@ fn raphtory(py: Python<'_>, m: &PyModule) -> PyResult<()> {
         global_temporal_three_node_motif_from_local,
         algorithm_module
     )?)?;
+    algorithm_module.add_function(wrap_pyfunction!(
+        all_local_motifs,
+        algorithm_module
+    )?)?;
 
     m.add_submodule(algorithm_module)?;
     let graph_loader_module = PyModule::new(py, "graph_loader")?;
