+++ raphtory/src/algorithms/motifs/mod.rs
@@ -1,2 +1,3 @@
-pub mod three_node_local;
+pub mod three_node_broken;
 pub mod three_node_motifs;
+pub mod three_node_local;
